removing superfluous code
