+++ csss-site/src/elections/views/utils/webform_to_json/transform_nominee_webform_to_json.py
@@ -12,7 +12,8 @@ def transform_nominee_webform_to_json(nominee):
     Keyword Arguments:
     nominee -- the dict that contains the nominee info that will be updated
     """
-    if ELECTION_JSON_KEY__NOM_POSITION_AND_SPEECH_PAIRINGS in nominee and type(nominee[ELECTION_JSON_KEY__NOM_POSITION_AND_SPEECH_PAIRINGS]) == dict:
+    remove_id_if_invalid(nominee)
+    if verify_that_position_and_speech_pairing_dict_is_in_nominee_dict(nominee):
         nominee[ELECTION_JSON_KEY__NOM_POSITION_AND_SPEECH_PAIRINGS] = list(
             nominee[ELECTION_JSON_KEY__NOM_POSITION_AND_SPEECH_PAIRINGS].values()
         )
@@ -40,6 +41,14 @@ def transform_nominee_webform_to_json(nominee):
                         }]
                     else:
                         position_and_speech_pairing[ELECTION_JSON_KEY__NOM_POSITION_NAMES] = [position_info[0]]
-            if ID_KEY in position_and_speech_pairing:
-                if (position_and_speech_pairing[ID_KEY] == "" or not f"{position_and_speech_pairing[ID_KEY]}".isdigit()):
-                    del position_and_speech_pairing[ID_KEY]
+            remove_id_if_invalid(position_and_speech_pairing)
+
+
+def remove_id_if_invalid(dict_obj):
+    if ID_KEY in dict_obj:
+        if dict_obj[ID_KEY] == "" or not f"{dict_obj[ID_KEY]}".isdigit():
+            del dict_obj[ID_KEY]
+
+
+def verify_that_position_and_speech_pairing_dict_is_in_nominee_dict(nominee):
+    return ELECTION_JSON_KEY__NOM_POSITION_AND_SPEECH_PAIRINGS in nominee and type(nominee[ELECTION_JSON_KEY__NOM_POSITION_AND_SPEECH_PAIRINGS]) == dict
